Type device routing with DeviceIdx enum class
Replace raw `uint8_t device_idx` (0=primary, 1=secondary, 255=none)
with `enum class DeviceIdx { Primary, Secondary, None }` throughout
the device routing layer. The previous untyped scheme made several
classes of bugs invisible at compile time, including the legacy 1/2
values left at SeqPage call sites by the prior slot->device_idx
migration. Those call sites were silently masked by a
`normalized_seq_idx` helper that mapped 2->0; this commit removes the
helper and remaps the remaining call sites to DeviceIdx::Primary or
DeviceIdx::Secondary as the original semantics required.

Scope:

- MidiDevice virtual interface: init_grid_devices, on_connection,
  disconnect, cleanup, add_track_to_grid take DeviceIdx. All driver
  subclasses (MD, A4, MNM, Generic, TBD) updated.
- DeviceContext::for_device, primary, secondary take DeviceIdx.
- DeviceParamTarget::device_idx field, DeviceParamResolver namespace
  functions (device_for_idx, target_count_for_idx, target_for_idx,
  perf_dest_from_idx) take DeviceIdx.
- DeviceManager::device_for_idx, context_for_device, attach_port,
  primary_context, secondary_context take DeviceIdx.
- SeqPage helpers (current_device_idx, select_device_idx,
  device_for_seq_idx, idx_is_md_device, seq_idx_for_device,
  seq_device_idx_name, opt_capture_is_md_device,
  opt_midi_device_idx_capture) take DeviceIdx.
- LFOSeqTrack::device_idx, MixerPage::mixer_device_idx,
  SeqPtcPage::last_midi_device_idx, TbdDevice::ui_device_idx_,
  TbdUiMode::device_idx_/bound_device_idx_ fields take DeviceIdx.
- LFOTrackRef, PerfPageParent, MidiActivePeering, TbdPanel callers
  switched to DeviceIdx values and comparisons.

Persisted storage stays uint8_t (mcl_cfg.seq_dev is menu-bound;
GridDeviceTrack::device_idx is on-disk project format). Boundary
casts are explicit via static_cast at those sites only.

The TbdUiMode::SLOT_PRIMARY/SECONDARY/NONE and TBD.cpp's
kTbdUiSlotPrimary/Secondary/None constants are removed in favor of
DeviceIdx values directly.

Both pio run -e tbd and pio run -e megacommand succeed.

+++ src/mcl/Drivers/DeviceCapabilities.cpp
@@ -11,7 +11,7 @@ DeviceMixerCapability::DeviceMixerCapability(MidiDevice &device)
     : DeviceCapability(device) {}
 
 uint8_t DeviceMixerCapability::track_count(const DeviceContext &ctx) const {
-  uint8_t grid_idx = ctx.device_idx();
+  uint8_t grid_idx = static_cast<uint8_t>(ctx.device_idx());
   if (grid_idx >= NUM_GRIDS) {
     return 0;
   }
@@ -29,7 +29,7 @@ uint8_t DeviceMixerCapability::track_count(const DeviceContext &ctx) const {
 
 SeqTrack *DeviceMixerCapability::seq_track(const DeviceContext &ctx,
                                            uint8_t track) {
-  uint8_t grid_idx = ctx.device_idx();
+  uint8_t grid_idx = static_cast<uint8_t>(ctx.device_idx());
   if (grid_idx >= NUM_GRIDS || track >= GRID_WIDTH) {
     return nullptr;
   }
